refactor(akgentic-frontend): 18-8 relocate FaviconService to core/config beside ConfigService
Move FaviconService from core/ui/ to core/config/ (it reads favicon/logo from ConfigService) and update imports; pure relocation, same root-provided singleton.

Relates to #158
