+++ src/app/app.component.spec.ts
@@ -14,7 +14,7 @@ import { ApiService } from './core/http/api.service';
 import { AuthService } from './core/auth/auth.service';
 import { ConfigService } from './core/config/config.service';
 import { ContextService } from './core/context/context.service';
-import { FaviconService } from './core/ui/favicon.service';
+import { FaviconService } from './core/config/favicon.service';
 import { ViewService } from './core/ui/view.service';
 
 function makeTeam(overrides: Partial<TeamContext> = {}): TeamContext {
