+++ src/app/core/config/favicon.service.ts
